.runpod/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
albucore==0.0.17
albumentations==1.4.18
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
audio-separator==0.21.2
audioread==3.0.1
av==12.1.0
beartype==0.18.5
boto3==1.35.41
botocore==1.35.41
braceexpand==0.1.7
cachetools==4.2.4
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contextlib2==21.6.0
contourpy==1.3.0
cpm-kernels==1.0.11
cycler==0.12.1
Cython==3.0.11
datasets==3.0.1
decorator==4.4.2
decord==0.6.0
deepspeed==0.14.4
diffq==0.2.4
dill==0.3.8
docker-pycreds==0.4.0
easydict==1.13
einops==0.8.0
eval_type_backport==0.2.0
executing==2.1.0
ffmpegcv==0.3.15
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==1.35.0
google-auth-oauthlib==0.4.6
hjson==3.1.0
huggingface-hub==0.25.2
humanfriendly==10.0
icecream==2.1.3
idna==3.10
imageio==2.34.2
imageio-ffmpeg==0.5.1
insightface==0.7.3
jax==0.4.36
jaxlib==0.4.36
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.5
lazy_loader==0.4
librosa==0.10.2.post1
lightning-utilities==0.11.8
llvmlite==0.43.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mediapipe==0.10.14
ml_collections==0.1.1
ml_dtypes==0.5.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.1
ninja==1.11.1.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.17.0
onnx2torch==1.5.15
onnxruntime-gpu==1.19.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
prettytable==3.11.0
proglog==0.1.10
propcache==0.2.0
protobuf==4.25.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyav==14.0.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.3
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
resampy==0.4.3
rotary-embedding-torch==0.6.5
rsa==4.9
s3transfer==0.10.3
safetensors==0.4.3
samplerate==0.1.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.0
sentencepiece==0.2.0
sentry-sdk==2.16.0
setproctitle==1.3.3
smmap==5.0.1
sounddevice==0.5.1
soundfile==0.12.1
soxr==0.5.0.post1
SwissArmyTransformer==0.4.12
sympy==1.13.3
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tifffile==2024.9.20
tokenizers==0.20.1
torch==2.4.0
torch-tb-profiler==0.4.3
torchmetrics==1.4.3
torchvision==0.19.0
tqdm==4.66.5
transformers==4.45.2
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.17.5
wcwidth==0.2.13
webdataset==0.2.100
xxhash==3.5.0
yarl==1.15.3
